+++ Modules/Intent.Modules.HotChocolate.GraphQL/Templates/MutationType/MutationTypeTemplatePartial.cs
@@ -37,6 +37,7 @@ namespace Intent.Modules.HotChocolate.GraphQL.Templates.MutationType
             AddTypeSource(TemplateFulfillingRoles.Domain.ValueObject);
             CSharpFile = new CSharpFile(this.GetNamespace(), this.GetFolderPath())
                 .AddUsing("System")
+                .AddUsing("System.Linq")
                 .AddUsing("System.Threading")
                 .AddUsing("System.Threading.Tasks")
                 .AddClass($"{Model.Name}", @class =>