+++ Modules/Intent.Modules.HotChocolate.GraphQL.Dispatch.Services/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.1
+
+- Fix : Collection conversion from `IReadOnlyList<T>` to `List<T>` on service dispatching.
+
 ### Version 1.0.0
 
 - Support for dispatching GraphQL queries and mutations to Application services through dependency injection.